# exp-lending-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/experience/lending/simulations/configure:
    post:
      tags:
      - Lending - Simulation & Eligibility
      summary: Configure Simulation
      description: Calculates a lending simulation via the pricing engine and persists
        the result through domain-lending-loan-origination.
      operationId: configureLendingSimulation
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigureSimulationCommand"
        required: true
      responses:
        "201":
          description: Simulation configured and persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfiguredSimulationDTO"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfiguredSimulationDTO"
  /api/v1/experience/lending/simulations:
    post:
      tags:
      - Lending - Simulation & Eligibility
      summary: Create Simulation
      description: "Computes a loan simulation for the given amount, term, and product\
        \ type."
      operationId: createSimulation
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSimulationCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResultDTO"
...
  /api/v1/experience/lending/eligibility:
    post:
      tags:
      - Lending - Simulation & Eligibility
      summary: Check Eligibility
      description: Evaluates whether a party is eligible for a given product and requested
        amount.
      operationId: checkEligibility
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckEligibilityCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResultDTO"
...
  /api/v1/experience/lending/simulations/{id}:
    get:
      tags:
      - Lending - Simulation & Eligibility
      summary: Get Simulation
      description: Retrieves a previously created loan simulation by its identifier.
      operationId: getSimulation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResultDTO"
  /api/v1/experience/lending/products/{productId}/pricing:
    get:
      tags:
      - Lending - Products
      summary: Get product pricing
      description: Retrieves the pricing configuration for a specific lending product.
      operationId: getLendingProductPricing
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Product pricing returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPricingDTO"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPricingDTO"
  /api/v1/experience/lending/products:
    get:
      tags:
      - Lending - Products
      summary: List products with pricing
      description: "Returns the lending products catalog with their pricing configuration,\
        \ optionally filtered by product type."
      operationId: listLendingProducts
      parameters:
      - name: productType
        in: query
        required: false
        schema:
          type: string
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Products returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductPricingDTO"
...
components:
  schemas:
...
    ConfigureSimulationCommand:
      required:
      - productType
      - requestedAmount
      - term
      type: object
      properties:
        productType:
          minLength: 1
          pattern: ^(PERSONAL_LOAN|LEASING)$
          type: string
        productId:
          type: string
          format: uuid
        requestedAmount:
          type: number
        term:
          minimum: 1
          type: integer
          format: int32
        purpose:
          type: string
        sector:
          type: string
        assetType:
          type: string
    ConfiguredSimulationDTO:
      type: object
      properties:
        simulationId:
          type: string
          format: uuid
        productType:
          type: string
        productId:
          type: string
          format: uuid
        requestedAmount:
          type: number
        term:
          type: integer
          format: int32
        monthlyPayment:
          type: number
        tin:
          type: number
        tae:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
    CreateSimulationCommand:
      type: object
      properties:
        amount:
          type: number
        term:
          type: integer
          format: int32
        productType:
          type: string
        productId:
          type: string
          format: uuid
    SimulationResultDTO:
      type: object
      properties:
        simulationId:
          type: string
          format: uuid
        monthlyPayment:
          type: number
        annualRate:
          type: number
        totalCost:
          type: number
        term:
          type: integer
          format: int32
        productType:
          type: string
...
    CheckEligibilityCommand:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        requestedAmount:
          type: number
    EligibilityResultDTO:
      type: object
      properties:
        evaluationId:
          type: string
          format: uuid
        eligible:
          type: boolean
        maxAmount:
          type: number
        reasons:
          type: array
          items:
            type: string
...
    FeeDefinitionDTO:
      type: object
      properties:
        type:
          type: string
        percentage:
          type: number
        fixed:
          type: number
    InterestRateBracketDTO:
      type: object
      properties:
        minAmount:
          type: number
        maxAmount:
          type: number
        tin:
          type: number
    ProductPricingDTO:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        productCode:
          type: string
        productType:
          type: string
        currency:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
        minTerm:
          type: integer
          format: int32
        maxTerm:
          type: integer
          format: int32
        interestRates:
          type: array
          items:
            $ref: "#/components/schemas/InterestRateBracketDTO"
        fees:
          type: array
          items:
            $ref: "#/components/schemas/FeeDefinitionDTO"
...